Replace windows (\r\n) and osx (\r) style newline encoding for the user agent of the rest client.
On Windows systems and legacy osx systems this would cause exceptions when trying to fetch information about maps or when trying to add maps from Trackmania Exchange. This is caused by the Mania Exchange or Trackmania Exchange Servers not accepting newlines in the User Agent Header of the Request.

EvoSC-127 Trackmania Exchange responding with 400: Bad Request